refactor(gr26): extract shelter ingest into new job/ package
Pulls the producer (TCMShelterBatch -> foreman enqueue) and the
consumer (foreman job -> S3 parquet decode) for gr26.ingest_latest_batch
out of service/ and into a new gr26/job/ package. service/ now only
holds CAN-frame business logic; new foreman job kinds belong in job/.

Files
  - gr26/job/shelter_batch.go (new) — owns both sides of the flow:
      * OnShelterBatchReceived: the producer-side hook
      * IngestLatestBatchHandler + processFile + dispatchRow: the
        worker-side ingest pipeline
      * shelterIngestParams, shelterRow: package-private types
      * replayFrame: own copy of decode-and-persist (no envelope, no
        upload-key check, no WS publish, no side-channel hook). Picked
        a standalone duplicate over a shared persistFrame so live and
        replay can diverge cleanly.

  - gr26/service/shelter_batch_hook.go    deleted
  - gr26/service/shelter_ingest.go        deleted
  - gr26/service/message.go               two small changes:
      * adds package-level var ShelterBatchHook (a func), called at
        the end of HandleMessage when canID == MsgIDShelterBatch. nil
        by default; main.go wires it to job.OnShelterBatchReceived.
        This is the seam that lets service stay agnostic of job
        (which depends on service) without an import cycle.
      * mustJSON -> MustJSON, exported so job.replayFrame can build
        the same status/note blobs without duplicating the helper.

  - gr26/main.go                          two new lines:
      * service.ShelterBatchHook = job.OnShelterBatchReceived (before
        mqtt.Init so the hook is set when the subscriber starts firing)
      * worker registry now points at job.IngestLatestBatchHandler

No behavior changes — same wire format, same foreman semantics, same
S3 path. Pure code organization.
