+++ gr26/main.go
@@ -6,6 +6,7 @@ import (
 	"github.com/gaucho-racing/mapache/gr26/api"
 	"github.com/gaucho-racing/mapache/gr26/config"
 	"github.com/gaucho-racing/mapache/gr26/database"
+	"github.com/gaucho-racing/mapache/gr26/job"
 	"github.com/gaucho-racing/mapache/gr26/mqtt"
 	"github.com/gaucho-racing/mapache/gr26/pkg/kerbecs"
 	"github.com/gaucho-racing/mapache/gr26/pkg/logger"
@@ -21,12 +22,18 @@ func main() {
 	config.PrintStartupBanner()
 	kerbecs.Init(config.KerbecsEndpoint, config.KerbecsUser, config.KerbecsPassword)
 	database.Init()
+
+	// Wire the service-level side-channel hook into the job module. Has
+	// to happen before MQTT.Init so the hook is set by the time the
+	// subscriber callback starts firing.
+	service.ShelterBatchHook = job.OnShelterBatchReceived
+
 	mqtt.Init(service.SubscribeTopics)
 
-	// Worker pool: foreman gives us jobs, we run them. Currently just
-	// gr26.ingest_latest_batch but adding more is a Register call.
+	// Worker pool: foreman gives us jobs, we run them. Each job kind
+	// lives in gr26/job/; register its handler here.
 	reg := worker.NewRegistry()
-	reg.Register("gr26.ingest_latest_batch", service.IngestLatestBatchHandler)
+	reg.Register("gr26.ingest_latest_batch", job.IngestLatestBatchHandler)
 	worker.StartPool(context.Background(), reg)
 
 	api.Run()
